Update rankedListGenerator.cpp
Adds a hopefully better method of quicksort2. Quicksort2 does not currently work and is a still work in progress to figure out why.
It is still a good idea, it adds spacial complexity, but should reduce the total number of comparisons.
Quicksort2's call in main is commented out because of this.